Added feeds handler implementation

+++ internal/store/feed.go
@@ -26,7 +26,7 @@ func CreateFeed(db *sql.DB, title string, url string, siteLink string) (int64, e
 }
 
 // GetFeed retrieves a feed by id from the database and stores it in the provided models.Feed struct.
-func GetFeed(db *sql.DB, id int) (*models.Feed, error) {
+func GetFeed(db *sql.DB, id int64) (*models.Feed, error) {
 	var f models.Feed
 	query := `SELECT id, title, url, site_link, created_at FROM feeds WHERE id = ?`
 	err := db.QueryRow(query, id).Scan(&f.ID, &f.Title, &f.URL, &f.SiteLink, &f.CreatedAt)
@@ -59,7 +59,7 @@ func GetAllFeeds(db *sql.DB) ([]*models.Feed, error) {
 }
 
 // DeleteFeed deletes a feed by id from the database.
-func DeleteFeed(db *sql.DB, id int) error {
+func DeleteFeed(db *sql.DB, id int64) error {
 	query := `DELETE FROM feeds WHERE id = ?`
 	_, err := db.Exec(query, id)
 	if err != nil {
